[wasm][cleanup] Change inlining from Wasm feature to V8 flag
Wasm-in-Wasm inlining used to be defined via a Wasm feature so that it
could be controlled in Origin Trials. As the optimization has shipped
and is enabled by default since May (https://crrev.com/c/5562486), we
should just make it a regular V8 flag.

This is more in-line with how we configure other optimizations, is
more obvious than the macros in `wasm-feature-flags.h`, removes the
(potentially confusing) experimental prefix from the command-line
flag, and complies with our policy to remove shipped feature flags
after two milestones (https://v8.dev/docs/wasm-shipping-checklist,
Wasm-in-Wasm inlining for all Wasm modules was enabled by default
in May 2024 by https://crrev.com/c/5562486).

Drive-by cleanup: We also used to have the (potentially confusing)
situation that Liftoff, Turbofan, and Turboshaft ignored the
`--experimental-wasm-inlining` flag for Wasm GC modules.
That is, inlining was always enabled for Wasm GC modules and could not
be turned off (except in Turboshaft and with `--no-liftoff`, which
is yet another special case to get rid off).
Since we inline for all Wasm modules by default, we should just
use the V8 flag to toggle the optimization consistently.
For production, nothing changes, and for testing and debugging, we can
now reliably turn inlining off via the flag.

Bug: 40898108
Change-Id: I85e4de09f5bd96340ca1a1208faebc670905c9c3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5882036
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96246}