[regexp] Cache results of RegExpPrototypeMatch
.. for a degenerate case in which the pattern is global and ATOM, and
the subject string is a SlicedString whose length grows with each
call.  In this case, we only need to search the 'new' part of the
substring.

This is a one-element cache and thus very cheap: a 4-elements
FixedArray, and no hashing/index calculations are involved.

Change-Id: I59628f31dd1ca89cefacf0a64b4c17be044465db
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5886727
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96271}