[foozzie] Enable dropping flags by content
This adds a new configuration of flags to drop when certain regular
expressions match on the content of the used test case.

The feature is enabled to drop --jitless when the %WasmStruct() or
%WasmArray() runtime functions are used.

The offending flag is dropped from the first two runs, which are a
baseline flavor and the default shipping configuration. It could
potentially be passed to these runs via extra flags generated by the
fuzzer.

The third run is dropped entirely when the flag is present, as it in
most cases would be the jitless experiment, which only contains the
--jitless flag. If it's dropped, the second and third run are
redundant.

This also adds a set of generic integration tests for testing the
feature functionality and two system tests specific to the
--jitless case.

No-Try: true
Bug: 368464066, 369105515
Change-Id: Ia5f3fecfd2259bc123f742a44cc52e4ed39a5dcc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5886726
Reviewed-by: Alexander Schulze <alexschulze@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96269}