[riscv] Enabling Maglev on RISC-V
Here is the implementation of  maglev compiler for RISC-V 64-bit
architecture.

We haven't implemented it for RV32, so we have disabled it in the
configuration options.

The main issue was the implicit usage of the flags register on all other
platforms supported by maglev (ARM, X86), where it was often generated
in one part and then used in another part (e.g., in JumpIf). On RISC-V,
we don't have such a register and cannot use it implicitly. Therefore,
we have introduced a pseudo flags register (s10), which is limited to
maglev. However, we tried to avoid using it wherever possible, so some
refactoring (encapsulation) of shared maglev code was necessary. This
refactoring does not affect the code executed (on other platforms); it
only affects where this code is located, allowing us to provide a custom
implementation for RISC-V.

The tests pass rate is the same with or without maglev on RISC-V without
compressed pointers. With compressed pointers, there are still a few
bugs to fix, and we are working on them. Static roots aren't supported so far,

We can see about +6% on Speedometer 2.0 on LicheePi4A (Speedometer 3
doesn't work there due to graphics driver issues) and some good
improvements in Octane/Kraken/Speedometer. This work has been done by
four people (including myself), all of whom have signed the CLA.

co-authors: Alexey Pavlyutkin <alexey.pavlyutkin@syntacore.com>
            Ilya Gavrilin <ilya.gavrilin@syntacore.com>
            Yuri Gaevsky <yuri.gaevsky@syntacore.com>

Bug: 342428026
Change-Id: Ie18cc0c469572be638637910d225b85d7fad8c75
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5729672
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Yahan Lu <yahan@iscas.ac.cn>
Reviewed-by: Yahan Lu <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#96295}