[inspector] Trace async parent per `v8::StackTrace`.
This fixes a long-standing inconsistency, where we'd capture synchronous
stack traces as `v8::StackTrace` (on the V8 heap) at some point, and
only associate the async stack trace when transforming this V8 internal
object into a `v8_inspector::V8StackTrace` instance. These two events
can happen at different points in time, and thus the async parent can
differ (and no longer match the correct async parent at the time when
the stack trace was captured). With this change we now use the
previously introduced ID on `v8::StackTrace` instances to memorize the
correct async parents in the `v8_inspector::V8Debugger`.

The inspector is notified of any stack trace capturing via the
`AsyncEventDelegate` interface, and at the point where a
`v8::StackTrace` is turned into a `v8_inspector::V8StackTrace` or a
`v8_inspector::AsyncStackTrace`, we look up the async parent based on
the ID of the `v8::StackTrace`. This way, async parents are now always
correctly matching the synchronous part of the stack trace reported from
the inspector.

The debugger will now eagerly create `v8::StackTrace`s whenever an error
object is created and the inspector is attached (and asking for detailed
stack traces upon uncaught exceptions). Prior to this change we tried to
delay the creation of the detailed stack trace for as long as possible,
but that optimization is likely overkill, since in Chrome the inspector
will almost always immediately ask for the detailed stack trace for
every exception thrown. But if this turns out to be a performance
problem in the real world (during debugging) we can always defer the
creation of the `v8::StackTrace` object a bit more.

Bug: 350426235
Change-Id: I4b299438ee38f04c4d8ef162e1a8c517e8d6ca0c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5882140
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96303}