[wasm][jspi] Fix eager freeing of unwinder stack memory
When a termination exception is propagated out of a JSPI export, the
unwinder returns the stack to the stack pool. However if the stack pool
is already full, the stack is immediately freed which invalidates the
memory that the unwinder is currently using.
Remove the excess stacks later, when we allocate a new one. This keeps
the memory alive long enough to finish unwinding, and still avoids
accumulating arbitrarily many stacks in the pool.

R=jkummerow@chromium.org

Bug: 42202153
Change-Id: I0a80302fe84e5fdf21f5f290b9c70af9b6bcfeb4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5891739
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96316}