[builtins] Add formal parameter count to CPP builtin definitions
... based on how SharedFunctionInfo objects using those builtins are
configured (function length value and whether arguments adaptation is
needed). In order to ensure that the formal parameter count values
were correctly transferred to CPP builtins definitions this CL adds
respective CHECKs to SharedFunctionInfo verification.

This is a step towards making the CPP builtins definitions the source
of truth for the formal parameter count value.

This CL temporarily introduces EmptyFunction1 builtin which does the
same as EmptyFunction (just returns undefined) but is configured
differently: it has formal parameter count 1 and requires arguments
adaptation. This is because EmptyFunction builtin was also used as an
implementation for some property setters which were set up as
JSFunctions with one argument requiring adaptation. Once the formal
parameter count specified in the builtins macro becomes the source of
truth this workaround can be removed.

Drive-by fixes:
 - make sure that SharedFunctionInfos with Illegal builtin are
   configured the same way (formal parameter count 0 and not requiring
   adaptation),
 - fix printing of SharedFunctionInfo's formal parameter count.

Bug: 366374966
Change-Id: I9cf0abe98ae42c8206db14b509736829b8e73474
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5899393
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96356}