[object] Change isolate retrieval in Name::GetRawHashFromForwardingTable
This CL moves the logic in HandleBase::IsDereferenceAllowed() to
compute the current isolate even for background threads into a new
method Isolate::CurrentMaybeBackground(). We can also use this method
then in Name::GetRawHashFromForwardingTable as a quick-fix to access
the isolate instead of GetIsolateFromWritableObject().

Bug: 336738728
Change-Id: I81c2a87d95dd13172b08a6192cb9dd97dad4aab9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5895530
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96359}