[wasm][js-api] Use BigInt for mem64 and table64
The JS API for WebAssembly.Memory and WebAssembly.Table objects should
accept and return BigInt instead of Number for 64-bit memories and
tables.

It's still being discussed if we want to allow anything that can be
converted to BigInt, or just BigInt itself, see
https://github.com/WebAssembly/memory64/issues/85.
For now, we only accept BigInt as currently spec'ed.

R=evih@chromium.org

Bug: 358381633
Change-Id: I5741aab5feec2a047320a355e3a5c5a0d4ccf673
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5898572
Reviewed-by: Eva Herencsárová <evih@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96362}