[class] deserialize receiver for instance member initializers
When there is an instance member initializer, an arrow function
in the computed field *key* can access super.prop, which after
serialization, can appear as if the class scope needs to access
the receiver of the outer scope. In this case, deserialize the
receiver if it appears to have one allocated.

Bug: 366863049
Change-Id: I3cf5cd46c135f75625b2fbbd34ac0f50fb2d9bc8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5872642
Commit-Queue: Joyee Cheung <joyee@igalia.com>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Joyee Cheung <joyee@igalia.com>
Cr-Commit-Position: refs/heads/main@{#96397}