[wasm][js-api] Handle exceptions more consistently
This introduces a WasmJSApiScope object which encapsulates what all our
callbacks need to do: Validating the FunctionCallbackInfo, opening a
HandleScope, allocating an ErrorThrower, providing access to the
v8::Isolate and i::Isolate.

This class also checks in its destructor that if we set a return value
then no exception is set on isolate. It provides a helper which can
manually be called to assert that an exception is or will be set on the
isolate on an error path.
This hopefully makes it easier to handle exceptions consistently in the
Wasm JS-API.

One case that was often handled wrong is to-string-conversions. For some
reason the external API (via `v8::Local<v8::String>`) swallows
exceptions. We now use the internal API (via `i::Handle<i::String>`),
which correctly propagates exceptions and also avoids a string
allocation because we can compare against a `base::Vector` directly.

R=jkummerow@chromium.org

Fixed: 370562605
Change-Id: I7483f3e32daf5bc52f18295460475f400e6339e5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5898794
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96394}