[riscv][wasm] Fixes EmitLazyCompileJumpSlot generation for C-extension
The first part of the problem is about Jump slot reservation: with
enabled C-extension the system reserves C.LD/C.LW instructions instead
of LD/LW for loading of Jump target address, and later it fails
patching the jump with real offset. The patch enforces generation of
exactly LD/LW instructions by stubbing the offset with 1 (cannot be
handled with C.LW/C.LD) instead of 0.

The second part of the problem relates to padding of generated Jump
slot to required size: it still may use C.LD/C.LW and so cannot be
padded to exactly kLazyCompileTableSlotSize with only a sequence of
NOP's. The patch fixes that with conditional insertion of C.NOP at the
end.

Change-Id: I4a153005dfd86c70decfeef24f24f460af5130d1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5907398
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Yahan Lu <yahan@iscas.ac.cn>
Commit-Queue: Yahan Lu <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#96459}