[etw] Fix missing builtins
After the patch that skips emitting the builtins, as their debugging
information is already added in the binary by the platform embedded
file writer, we observed the interpreter trampoline builtins were
missing.

This is because they are also builtins, but we do not want to skip
them as otherwise the stack traces is incomplete.

We also observed that even builtins in DLL can be missing if they
are relocated.

Bug: 42202931
Change-Id: I72bd31219b0c1366bd5e9e295592ccfeb64912d2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5905047
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: José Dapena Paz <jdapena@igalia.com>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96696}