[sandbox][leaptiering] Validate argument count for optimized JS calls
Compilers can optimize calls to JSFunctions if the call target is known
at compiler time. In that case, they will not use the code new code path
introduced in crrev.com/c/5832362 which ensures that the correct
parameter count is used. As such, with this CL we now emit a short call
prologue that guarantees that the correct parameter count is used, even
if the JSFunction object is corrupted.

In the future, we'll want to simply hardcode the JSDispatchEntry into
the generated code and use the parameter count from that entry during
compilation. Then we can statically guarantee that the parameter counts
match without any additional verification at runtime. However, this
first requires the GC to be able to track dispatch entries referenced
from generated code.

Bug: 40931165, 42204201, 342297062
Change-Id: Ibc31e5f449cc4b2e889b56b62cb54e983a636b15
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5844709
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96406}