[sandbox] Define kJavaScriptCallDispatchHandleRegister
We want to change the JS calling convention so that the caller passes
the dispatch handle to the callee. This will for example allow the
callee to determine the total number of arguments (which depends on the
formal parameter count of the invoked function). This CL is a first step
in that direction: we define a new kJavaScriptCallDispatchHandleRegister
which the caller must set to the dispatch handle of the called function.

Bug: 40931165
Change-Id: I6534143e9c4d5042ed84c869f1a0812e7cc5d765
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5857400
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96512}