Fix regression in instruction selection speed
The fix for 356196918 caused a dramatic slowdown when compiling certain
very large functions. This is because each call to
ZeroExtendsWord32ToWord64 did work proportional to the size of the
function, and the number of calls could also be proportional to the size
of the function. In this updated fix, rather than zeroing the entire
phi_states_ array, we correct the incorrect values in that array and
leave them for subsequent calls to ZeroExtendsWord32ToWord64. This
change also lazily allocates the phi_states_ array and adds a regression
test for 356196918.

Bug: 369883716, 356196918
Change-Id: I184e789d20d12863cd84e4474f857c56a22ce71f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5906045
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#96544}