[wasm][test] Add first tests for use counters
This adds basic tests for Wasm use counters, for all compilation APIs
(sync, async, streaming).
We do not cover lazy validation yet, which is disabled by default
anyway. If it is ever enabled or tested by default, the tests will fail
in that variant.

Fix sync compilation to always report the use counters, even if the
module comes from the cache. This makes us independent of GC timing, and
is in line with async and streaming compilation.

Plus some drive-by simplifications in tests due to a new EmitCode method
on the WasmFunctionBuilder.

R=jkummerow@chromium.org

Bug: 368070338
Change-Id: Ia41fc30738783763bd04a14c63fa2703723b8c73
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5904099
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96412}