[GC] Unaccount for external string memory explicitly
This CL adds explicit/manual calls to unaccount for external string
memory.
This allows us to bail out for strings that are owner by the main isolate: the accounting is already broken because we account for
in the isolate that creates the string and unaccount in the main one.
Instead of fixing, this CL avoids accounting altogether for those strings on both side (creation and deletion).

Follow-up will override Unaccount() in blink string.

Bug: chromium:371050124
Change-Id: Id6d76cdc04e3884b01a70ad71cbf63d660741662
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5905826
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96493}