[riscv] Add B-Extension detection
This change implements the detection of bit manipulation extensions
(Zba, Zbb, Zbs) using the hwprobe call. The hwprobe call requires a
glibc version of 2.39 or higher. Therefore, if the glibc version is
lower, the bit manipulation extensions can still be enabled manually
by using the riscv-b-extension flag. It is important to note that the
riscv-b-extension flag forces the activation of these extensions, so
it should be used with caution and respect.

Change-Id: I0163275a571c96fe877ddba10b7f41ab9ab44aaa
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5912821
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96652}