[turboshaft] Do not duplicate SwitchOp targets with a single GotoOp
If a `SwitchOp` has multiple cases with the same target, these result
in different target blocks in the CFG that then have a single
`GotoOp`. We should not clone any blocks into these "dummy" blocks as
blocks will not be used in the lowered jump table but the actual
(shared) target block instead.

Change-Id: I303f9cb001764116405342486167053395da6a4c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5912822
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96464}