[revec][wasm][turboshaft] Preserve valid Simd128 store in OOB trap
When the first SIMD128 store is in-bound while the second store is
out-of-bound, it's possible to miss a valid store after revectorization.

This patch is to always preserve the first lower Simd128 store to ensure
a consistent memory state before and after revectorization.

Bug: v8:12716
Change-Id: I7ed972da3b9637128d2549ad96cf287acb086512
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5880861
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Yolanda Chen <yolanda.chen@intel.com>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96461}