[execution] GetIsolateForSandbox() now returns IsolateForSandbox
Retrieving an isolate pointer from shared objects is problematic
on worker isolates because it always returns the shared space isolate
and not the current worker isolate.

Restricting the return type of GetIsolateForSandbox() to the newly
introduced IsolateForSandbox helps in ensuring that this method is
safe to call for shared objects. IsolateForSandbox only allows
accessing the external pointer table and for shared objects it
does not matter for which isolate the shared external pointer
table is accessed because it is shared across all main and worker
isolates.

Bug: 336738728
Change-Id: I06a051f15ac9135b4f928843db8893354830684d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5913523
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96471}