[cleanup] Unmerge array searching
The FixedArray header had templated functions for searching a sorted
array, templated into a split between an "all entries" search and a
"valid entries" one.

In practice, this was effectively a split into TransitionArray and
DescriptorArray searches, where the two arrays are sufficiently
different (TransitionArray is directly sorted, while DescriptorArray
keeps a separate set of sorted indices), and the use cases sufficiently
different (for TransitionArray we wanted an insertion index, for
DescriptorArray we wanted to limit the max index) that the
generalisation across the template split became a bit messy, and
required non-obvious behaviour to make the two arrays compatible with
the search interface.

Binary/linear search are reasonably easy to implement, particularly
using std::lower_bound, so we can de-abstract this search code and
duplicate it for the two use cases, making each copy simpler and
specialised for the custom behaviour it needs.

Change-Id: I480a4578b47a64e9d66e23be80ccee1f8a59269f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5913527
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96474}