[deserializer] Check source hash before starting BG merge
When a background deserialize task receives the source text, it looks up
that source in the isolate cache and prepares a background merge with
the matching in-memory cached script. However, we might have a source
mismatch between the source text and the serialized data; in this case,
the background merge will end up trying to merge incompatible scripts
and crashing.

Since we only start the background merge when the source text is
available, we can also at this moment sanity check the source hash. This
is a weak hash, so it won't prevent all issues, but it does at least
make this code more robust.

Change-Id: I3a5792755747e3d22cf1d75aa650c97a7f429de1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5920444
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96534}