[codegen][arm64] Use a zone to track unresolved branches
Use the zone allocator to keep track on unresolved branches, via a new
`ZoneAbslBTreeMap` wrapper for `absl::btree_map`. To do this, we need
assembler users to pass down a zone.

If a zone is available, it should be passed down to the macro-assembler
or assembler. Otherwise, assemblers need at least an
`AccountingAllocator` to create a local zone.

To do this, we introduce a `MaybeAssemblerZone zone` parameter to all
assemblers, which is created implicitly from either a `Zone*` or
`AccountingAllocator*`. Then only the arm64 assembler uses it, and
optionally creates a local zone if only an accounting allocator is
present.

For compatibility, given the Zone is only used on Arm64, we keep the old
constructors around for non-Arm64 architecture-specific code.

Bug: 347266976
Change-Id: I6484ee96275ff5a7e675baa25ed7704edcf11cdd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5713169
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Pierre Langlois <pierre.langlois@arm.com>
Cr-Commit-Position: refs/heads/main@{#96515}