[wasm] Skip spec tests which are unchanged in proposal repos
Instead of only checking if a test file is identical to the version in
the spec repo, do also check if there was any change in the proposal
repo since it branched from the spec repo.
This skips tests which have been updated in the spec repo, but not in
the proposal repo. The version in the proposal repo then often fails
until the proposal is rebased.

R=jkummerow@chromium.org

Change-Id: I6aa82a87822533bf1cd81c6d36f834092e746ba2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5920423
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96518}