[turboshaft] Allow multiple Gotos to the same loop header
The invariant that loops should have a single incoming forward edge
still stands.

However, now, when multiple Gotos are emitted and are going to the
same loop header, the Assembler takes care of introducing a new
"single_loop_predecessor" block that acts as the single forward
predecessor for the loop, and where all predecessors are routed to.

This is needed in particular when we have a graph like this:


                       |      |-------------------------|
                       |      |                         |
                       |      |                         |
                  +-- loop 1 ----+                      |
                  | ...          |                      |
                  | Goto         |                      |
                  +--------------+                      |
                          |                             |
                          |                             |
                   +----------------+                   |
                   | Call           |                   |
                   | CheckException |                   |
                   +----------------+                   |
                     /         \                        |
                    /           \                       |
                   /             \                      |
      +-----------------+      +------------------+     |
      | DidntThrow      |      | CatchBlockBegin  |     |
      | ....            |      | ...              |     |
      | Goto            |      | Goto             |     |
      +-----------------+      +------------------+     |
             |                          |               |
             |                          |               |
             |     |---------|          |   backedge    |
             |     |         |          |---------------|
             |     |         |
      +-- loop 2 -------+    |
      | ...             |    |
      | Goto            |    |
      +-----------------+    |
             |               |
             |               |
             |   backedge    |
             |---------------|


Here, as far as Turboshaft is concerned, loop 1 doesn't have any inner
loops, because loop 2 is an infinite loop that never goes back to the
header of loop 1 (actually, it doesn't even have to be an infinite
loop: it could also just contain a return). So, Turboshaft can unroll
loop 1.

The other thing that come into play is how CheckException / DidntThrow
/ CatchBlockBegin are processed by CopyingPhase. In
AssemblerOutputGraphCheckException, the CopyingPhase emits the
CheckException, then Binds the DidntThrow block and inlines the old
DidntThrow block in it (cf
https://source.chromium.org/chromium/chromium/src/+/main:v8/src/compiler/turboshaft/copying-phase.h;l=908-910;drc=618cea5a4ba57d7442880d20000628389da34e55).

So, when unrolling loop 1 in the drawing above, we end up duplicating
the DidntThrow block n times (where n is the number of time we unroll
the loop), and all of them produce a Goto going to the loop header of
loop 2.

So far, because of how graphs are built, it seems that this problem
only affects graphs built from Maglev. However, I wouldn't be
surprised if Wasm graphs or JS graphs coming from Turbofan end up
at some point being subject to this bug.

Fixed: chromium:358461173
Bug: chomium:42204525
Change-Id: I93afef0c0642c83ede142446037f27da73990f40
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5912940
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96527}