Reland "[common] Introduce macros for faster TLS usage"
This is a reland of commit 9baf91c6b8d9684f7f7405d2771ec8cd5140f755

This CL was reverted because of a build failure on ChromeOS. Fix
this by using "local-dynamic" for ChromeOS at the moment. This is still
better than the default of "global dynamic" on tip-of-tree and
hopefully we can switch to a faster model once we understand why
exo_wlcs uses V8.

Original change's description:
> [common] Introduce macros for faster TLS usage
>
> Introduce macros that choose the right TLS mode depending on the
> build mode similar to renderer/platform/heap/thread_local.h in Blink.
>
> This CL also then uses these macros on g_current_isolate_ and
> g_current_local_heap_.
>
> Bug: 336738728
> Change-Id: Ia97cf965dc76c2cb3af59f945036f2a7b105fc3a
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5920088
> Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
> Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#96520}

Bug: 336738728
Change-Id: I7f83339829e514a9a89ca4d457d5877a363b28d6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5922879
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96589}