[codegen] Don't pass a nullptr in InitUnwindingRecord for x64
After CL https://crrev.com/c/5713169, MacroAssemblerBase constructor
expects isolate to be non-null, but InitUnwindingRecord for x64 is
passing a nullptr to the MacroAssembler constructor. This breaks
mksnapshot for Windows debug builds with v8 symbol level 2.

Updated the code with the logic that is applied to the arm64 version
of the same function.

Change-Id: Id8a6a292e62d2e9f70011dafc562659bb2f3d178
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5924806
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Luis Pardo <lpardosixtos@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#96585}