Reland "[fastapi] Handle returns without type conversion to tagged"
This is a reland of commit 198db8e5fbc7d9719e53baf025aeadf0a8b5a441

This is the original CL without changes.

The original CL was reverted because of a performance regression
in motionmark/Leaves. The performance regression was probably caused
by a code alignment issue, or something similar. The regression was
caused by an `__ Unreachable()` that got introduced in this CL, which
did not influence any optimizations, but just made a code block of 11
instructions disappear.

Motionmark performance owners agreed to reland this CL, as it also
improves motionmark/Canvas Lines, and overall the motionmark score
does not change.

Original change's description:
> [fastapi] Handle returns without type conversion to tagged
>
> So far, returns of fast API calls got converted unconditionally to
> tagged. This meant that if the return type was float64, and the calling
> function could handle a float64, there would still be a HeapNumber
> allocation.
>
> With this CL, the result of the fast API call is marked with the correct
> type, so that e.g. float64 returns only get converted to HeapNumbers
> when necessary.
>
> Bug: 41497260
> Change-Id: Ib027e2dbacf5ac12baba6f64d7bd45f99c5b95d2
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5832361
> Commit-Queue: Andreas Haas <ahaas@chromium.org>
> Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#96032}

Bug: 41497260
Change-Id: Ic975f2f9e89e7e76e7eec1fdaa5fc52ae2a9c2bb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5892799
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96563}