[objects] Pass isolate as argument to NormalizedMapCache methods
This CL adds an isolate argument to various methods in
NormalizedMapCache and Map::Hash(). The callers of those methods
all have the isolate available already. There is no need to recompute
it again.

The GetIsolateFromWritableObject was triggering a DCHECK failure on
shared objects.

Bug: 372819448
Change-Id: I738378f6d55d7ab5d0071eb17990a910fb93eaf4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5920242
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96561}