[base] Clean up Vector construction and conversion
Allow converting `Vector<T>` and Vector<const T>` to `Vector<const U>`
if `T*` converts to `const U*` and `T` and `U` have the same size.

As long as we do not write to the `U` pointer, this is safe. And we
cannot write because we only allow this for `const U*`.

Also audit Vector construction and remove unneeded casts and redundant
template arguments, sometimes by switching to `VectorOf`.

R=jkummerow@chromium.org

Change-Id: I76cfa7328ac5adc210fbe1c662716968ca58ecf8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5913528
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96566}