[wasm-gc] Also run WasmGCOptimizePhase on typed-funcref functions
A function doesn't have to use an instruction from the (original)
wasm-gc to benefit from the WasmGCOptimizePhase, as it also optimizes
the instructions from the typed-funcrefs proposal.

This is really just an implementation detail and it isn't expected at
all to be relevant in real-world code but it creates weird corner cases
where reproducers need to add dummy instructions with the wasm-gc
prefix to enforce running the WasmGCOptimizePhase in turboshaft.

Bug: 42204049
Change-Id: I7a5b35c5803ce65de1933652e83d3a84641776e1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5929030
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96582}