[parser] Simplify variable lookup caching
Always use the first deserialized scope as cache. Allow function vars to
be "redeclared" if we don't find it in the cache, but simply return an
already deserialized function_ if it exists on redeclaration. This can
happen since variable conflict detection needs to cache deserialized
variables differently due to catch scopes: catch scopes don't generate
redeclaration contexts, but we still may have a redeclaration conflict
with a variable behind the catch context, so this kind of resolution
looks beyond a "masking variable". That's different from regular scope
resolution; and hence the lookup results cannot be reused.

Bug: 370694832
Change-Id: I1edbd81d1fbab13f97437fa75567c87a6e18b562
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5929088
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96595}