[string] Use an unsigned integer for String::length
With the sandbox, it is assumed that an attacker can corrupt heap
memory, and therefore String objects and in particular their lengths. If
the string lengths is represented as a signed integer (as it used to be
before this CL), there is some risk of signed integer comparisons when
operating on them as an attacker could turn the length into a negative
value. See https://crbug.com/327473424 for an example. This CL now
converts the length to a uint32_t to better defend against such issues.

Bug: 369748453
Change-Id: Ife1fc2a1e2c23ff022dee1c5e315650c95057aa3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5332375
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96599}