[float16array] Throw in Atomics methods
This CL fixes Atomics methods to throw when operating on Float16Arrays.

It does this by reordering the TypedArray ElementsKinds so that all
valid-for-atomic-access TA ElementsKinds are in a contiguous range that
comes before invalid-for-atomic-access ElementsKinds.

This CL also adds additional static and runtime safety checks.

Finally, drive-by removal of some dead methods.

Fixed: 372932101,373467913
Change-Id: I0bc72c403c225c44bcbf9bb59558c78039b2573f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5935819
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96638}