[sandbox] Tighten types for deopt trusted objects
DeoptimizationData is a ProtectedFixedArray, which means its elements
should be TrustedObject (or Smi). This CL fixes the element type of
ProtectedArrayShape to be a Union of Smi and TrustedObject (like the
comment used to say), rather than just Object.

This exposed the fact that TrustedWeakFixedArray was not in fact a
TrustedObject(!), and required tightening the type of the
SharedFunctionInfoWrapper field to SharedFunctionInfoWrapperOrSmi.

Additionally we can tighten the return value of
DeoptimizationData::SharedFunctionInfo (renamed to GetSharedFunctionInfo
to avoid the name clashing with the type) to SharedFunctionInfo instead
of Object, and remove some unnecessary casts.

Bug: 336140976
Change-Id: I44c8e5ca69f0edd274fa4a36eb83f63d9cc4e427
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5938027
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96645}