Add LookupCodeAndOffset helper function
We see some crashes in the wild in *ExitFrame::Summarize, trying to
authenticate the PAC signature of the frame's pc for calculating the
pc offset.
Introduce a LookupCodeAndOffset helper function that only reads the pc
once (LookupCode is already using maybe_unauthenticated_pc) and use that
for offset calculations.

Bug: 373677811
Change-Id: I983d0730d85847ae33c0d158cf18accc5fd9ea2e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5938029
Commit-Queue: Stephen Röttger <sroettger@google.com>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96671}