[sandbox][leaptiering] Pass dispatch_handle during JS function calls
Some callees need to be able to determine the number of arguments passed
to them. Due to argument padding, this requires knowledge of both the
actual argument count and of the formal parameter count of the invoked
function (the number of arguments on the stack is then the maximum of
the two). With the sandbox attacker model, it is not safe to load the
formal parameter count from within the sandbox. Instead, this CL adds a
new parameter to the JS calling convention which contains the
JSDispatchHandle that was used to invoke a function. The callee can then
simply use that dispatch handle to load the (trusted) parameter count
from the JSDispatchTable.

Drive-By: add various DCHECKs and static_asserts into places that rely
on the JavaScript calling convention.

Bug: 40931165, 42204201
Change-Id: I009606c298af2071052c859bfc8b27c5fcc72072
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5907404
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96672}