[parser] Don't handle `this` var management in reparsed scopes
The variable is already properly allocated, so we shouldn't try to
reallocate it. Notable this would be wrong in the case where we reparse
a class body for one of its initializers. The function in which the
class body actually lives might not have required a context, but
computed property names might have referred to its receiver. When we
reparse the class that function might not be on the outer scope chain of
the class (since it didn't require a context).

Bug: 371237564
Change-Id: I54ccbc86eb9abdcd558d395f7896d1a23a110b50
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5937959
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Joyee Cheung <joyee@igalia.com>
Cr-Commit-Position: refs/heads/main@{#96677}