[fastapi][test] Remove incorrect fuzzing code
Fast api call targets and regular api call targets are supposed to
behave the same, e.g. if the fast call target throws, the regular call
target is supposed to throw for the same input.

For one test api this was not the case when the fuzzing flag was set.
This CL removes the special handling of the fuzzing flag. The special
handling originally guarded a CHECK, but this CHECK has been removed in
the meantime, and the special handling is not needed anymore.

R=marja@chromium.org

Bug: 373719189
Change-Id: I29a59499f914dd03ecd2742e9478e594148acddd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5941624
Reviewed-by: Marja Hölttä <marja@chromium.org>
Auto-Submit: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96684}