[wasm] CanonicalValueType: finishing touches
This CL addresses most of the remaining TODOs, drops the implicit
conversions and non-type-strict comparisons, and reshuffles methods
in value-type.h to appropriate access levels (making most of the
superclasses "protected" instead of "public").
This flushed out the fact that class CanonicalValueType was
accidentally deriving from ValueType rather than ValueTypeBase,
which hid a few necessary adaptations in code elsewhere, mostly in
the compilers. That in turn required a solution for determining
whether a CanonicalValueType should use WasmNull or the JS NullValue,
which is now the only non-trivial change in this CL.

Bug: 366180605
Change-Id: I7c56f767c88b7c8cd41533171fd3557860c84e07
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5940189
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96688}