[heap] Fix exception in measureMemory in d8
The default MeasureMemoryDelegate was using V8 internal types. This is
not okay for the promise as it's .then() may be hooked by user JS.

In such a case we wouldn't have entered V8 properly which is a problem
when dealing with exceptions.

The fix uses the proper API types for context and promise. Following
up here we should just deprecate the V8 API and implement the delegate
in d8 with just using API types.

Bug: 371237560, 374253377
Change-Id: Iecc8553fd2bad4104162299571fdfa93ef171843
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5942942
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96718}